Попробовал разделить компонент Инпута на умный и глупый
