+++ src/app/styles/styles.scss
@@ -45,7 +45,6 @@ body {
 .Sidebar {
   border-radius: 0 0 0 10px;
   background-color: #fff;
-  padding: 24px;
   border-right: 2px solid #e6e6e6;
   grid-area: Sidebar;
 }
