+++ src/components/MessageInput/styles.scss
@@ -39,3 +39,7 @@
   top: 22px;
   left: 12px;
 }
+
+.ppp {
+  background-color: red;
+}
