+++ src/components/index.js
@@ -1,4 +1,4 @@
 export * from "./Message";
 export * from "./MessageBox";
-export * from "./MessageInput";
+export * from "./MessageInput/messageInputContainer";
 export * from "./MessageList";
